flask-backend/requirements.txt
flask
torch
Pillow
werkzeug